requirements.txt
fastapi
uvicorn
passlib
python-dotenv
psycopg2
jose